+++ NEWS.md
@@ -1,10 +1,22 @@
+## SWMPr 2.4.2
+
+* Added `fill` argument to `plot_summary` as option to fill missing values with long term monthly averages or linear interpolation
+
+* Added `base_size` argument for global text size in `plot_summary`
+
+* Fix to annual anomalies in `plot_summary` as average of monthly sums
+
+* Enhancement to `qaqc` to work if angle brackets not included in qualifier codes
+
+* Only required functions from data.table dependency imported
+
 ## SWMPr 2.4.1
 
 * Fix to `map_reserve` function for updates to ggmap
 
 * Added Pacific/Honolulu time zone and `'hee'` site code to `time_vec` function for He'eia reserve
 
-* Added `colsleft`, `colsmid`, and `colsright` arguments to `pot_summary` for separate colors on left, mid, right portions of plot
+* Added `colsleft`, `colsmid`, and `colsright` arguments to `plot_summary` for separate colors on left, mid, right portions of plot
 
 * URL fix to `metab_day` documentation